move carbon footprint text to dedicated card
https://github.com/e-mission/e-mission-docs/issues/961#issuecomment-1707597185 -> implementing this suggestion

isolate the text to a dedicated card, and place the "meter" card and "text" card in a carousel, now we have three rows, each a carousel.

also isolated data management functions shared across the two cards into `metricsHelper.ts`

The two cards keeps information easily accessible to those using a screen reader, while maintaining focus on the "meter" card and not cluttering the screen
